+++ CHANGELOG.md
@@ -127,6 +127,7 @@ changes.
 - Changed documents to prepare for open source [Issue 737](https://github.com/IntersectMBO/govtool/issues/737)
 - Changed copy on maintenance page [Issue 753](https://github.com/IntersectMBO/govtool/issues/753)
 - Update link to docs [Issue 1246](https://github.com/IntersectMBO/govtool/issues/1246)
+- Change error modal while connecting the wallet that has no supported extensions [Issue 1141](https://github.com/IntersectMBO/govtool/issues/1141)
 
 ### Removed